Fixed a matrix precision bug
In the case where a matrix had a value that was
smaller than the precision of a 32 bit float, we
saw an error on chrome. This commit guarantees
that exponents like 1e-88 are not output by
the matrix function.

+++ src/matrix.js
@@ -166,7 +166,11 @@ SVG.Matrix = SVG.invent({
     }
     // Convert matrix to string
   , toString: function() {
-      return 'matrix(' + this.a + ',' + this.b + ',' + this.c + ',' + this.d + ',' + this.e + ',' + this.f + ')'
+
+      // Construct the matrix directly
+      return 'matrix(' + float32(this.a) + ',' + float32(this.b) + ','
+        + float32(this.c) + ',' + float32(this.d) + ',' + float32(this.e)
+        + ',' + float32(this.f) + ')'
     }
   }
